[Design]DaangnMarket#69 - 채팅뷰 왼쪽 네비바
