+++ DaangnMarket/DaangnMarket/Presentation/Chatting/ViewController/ChatViewController.swift
@@ -1,0 +1,45 @@
+//
+//  ChatViewController.swift
+//  DaangnMarket
+//
+//  Created by 김민 on 2023/04/16.
+//
+
+import UIKit
+
+class ChatViewController: UIViewController {
+
+    // MARK: - View Life Cycle
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        setBackgroundColor()
+        setNavigationBar()
+    }
+}
+
+extension ChatViewController {
+    
+    // MARK: - Custom Methods
+    
+    private func setBackgroundColor() {
+        view.backgroundColor = .white
+    }
+    
+    private func setNavigationBar() {
+        navigationLeftBar()
+        navigationRightBar()
+    }
+    
+    private func navigationLeftBar() {
+        let titleLabel = UILabel()
+        titleLabel.text = "채팅"
+        titleLabel.font = UIFont.systemFont(ofSize: 20, weight: .bold)
+        self.navigationItem.leftBarButtonItem = UIBarButtonItem(customView: titleLabel)
+    }
+    
+    private func navigationRightBar() {
+        
+    }
+}
